make WARCRecordBuilder an interface
this way other classes that extend other classes can also implement
WARCRecordBuilder
